+++ src/repl/scala/tools/nsc/Interpreter.scala
