+++ src/repl/scala/tools/nsc/InterpreterLoop.scala
