+++ src/repl/scala/tools/nsc/interpreter/AbstractOrMissingHandler.scala
