+++ src/repl/scala/tools/nsc/interpreter/CommandLine.scala
