+++ src/repl/scala/tools/nsc/interpreter/Completion.scala
