+++ src/repl/scala/tools/nsc/interpreter/CompletionAware.scala
