+++ src/repl/scala/tools/nsc/interpreter/CompletionOutput.scala
