+++ src/repl/scala/tools/nsc/interpreter/ConsoleReaderHelper.scala
