+++ src/repl/scala/tools/nsc/interpreter/Delimited.scala
