+++ src/repl/scala/tools/nsc/interpreter/Formatting.scala
