+++ src/repl/scala/tools/nsc/interpreter/ISettings.scala
