+++ src/repl/scala/tools/nsc/interpreter/Imports.scala
