+++ src/repl/scala/tools/nsc/interpreter/InteractiveReader.scala
