+++ src/repl/scala/tools/nsc/interpreter/JLineCompletion.scala
