+++ src/repl/scala/tools/nsc/interpreter/JLineReader.scala
