+++ src/repl/scala/tools/nsc/interpreter/Logger.scala
