+++ src/repl/scala/tools/nsc/interpreter/LoopCommands.scala
