+++ src/repl/scala/tools/nsc/interpreter/MemberHandlers.scala
