+++ src/repl/scala/tools/nsc/interpreter/Naming.scala
