+++ src/repl/scala/tools/nsc/interpreter/Pasted.scala
