+++ src/repl/scala/tools/nsc/interpreter/Power.scala
