+++ src/repl/scala/tools/nsc/interpreter/ReplConfig.scala
