+++ src/repl/scala/tools/nsc/interpreter/ReplProps.scala
