+++ src/repl/scala/tools/nsc/interpreter/ReplReporter.scala
