+++ src/repl/scala/tools/nsc/interpreter/ReplStrings.scala
