+++ src/repl/scala/tools/nsc/interpreter/ReplVals.scala
