+++ src/repl/scala/tools/nsc/interpreter/Results.scala
