+++ src/repl/scala/tools/nsc/interpreter/RichClass.scala
