+++ src/repl/scala/tools/nsc/interpreter/SimpleReader.scala
