+++ src/repl/scala/tools/nsc/interpreter/package.scala
