+++ src/repl/scala/tools/nsc/interpreter/session/FileBackedHistory.scala
