+++ src/repl/scala/tools/nsc/interpreter/session/History.scala
