+++ src/repl/scala/tools/nsc/interpreter/session/JLineHistory.scala
