+++ src/repl/scala/tools/nsc/interpreter/session/SimpleHistory.scala
