+++ src/repl/scala/tools/nsc/interpreter/session/package.scala
